lesson15-qphan part-02

+++ lesson15-qphan/src/bean/Item.java
@@ -9,22 +9,22 @@ import java.util.Objects;
  */
 public class Item {
 
-	private Integer id;
+	private Long id;
 	private String name;
 
 	public Item() {
 	}
 
-	public Item(Integer id, String name) {
+	public Item(Long id, String name) {
 		this.id = id;
 		this.name = name;
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
